requirements.txt
flask
apscheduler
requests
pywebpush
waitress
pystray
Pillow
